Extract file watcher from codex-core
Move the file watcher implementation and tests into a dedicated codex-file-watcher crate while preserving codex_core::file_watcher as a compatibility re-export.
