+++ codex-rs/file-watcher/BUILD.bazel
@@ -1,0 +1,6 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "file-watcher",
+    crate_name = "codex_file_watcher",
+)
