+++ codex-rs/file-watcher/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "codex-file-watcher"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lib]
+name = "codex_file_watcher"
+path = "src/lib.rs"
+
+[lints]
+workspace = true
+
+[dependencies]
+notify = { workspace = true }
+tokio = { workspace = true, features = [
+    "macros",
+    "rt",
+    "sync",
+    "time",
+] }
+tracing = { workspace = true, features = ["log"] }
+
+[dev-dependencies]
+pretty_assertions = { workspace = true }
+tempfile = { workspace = true }
